Making the default target just the static library.
Additionally, abstracting the build process slightly more. The goal is
to make the library build and be usable without loadpng nor dependencies.
Now the library should be the default build. If users want, they can build
the example program with the `example` target or build everything with the
`all` target.